allow project-name config to be set via COMPOSE_PROJECT_NAME
the project-name config ends up in COMPOSE_PROJECT_NAME in the end. This change at least considers it as a way to set/keep it. The order of precedence is for setting project-name is (lowest to highest):

* read from .lc.yml
* read from COMPOSE_PROJECT_NAME
* read the value passed via the --project-name argument

if all are blank, falls back to using the cwd as the project-name