clean up failure handling
originally, I was afraid panicing in a command would circumvent an App's After hook. turns out after is called with defer
https://github.com/codegangsta/cli/blob/master/app.go#L258

I still like the pattern of commands returning errors. This change converts commands returning errors into a panic. It does this by wrapping the command when adding it to the App