refactor dev-env to allow GOPATH setup on mac
Unfortunately it is not easy to share the entire GOPATH with the mac OS, because mounting '-v $GOPATH:/go' will cause the mac os x GOPATH to override the container's GOPATH (note: I did not attempt to figure out how to have dueling GOPATHS in the container). This commit works around this by forcing some mucking with the local GOPATH.

I've refactored the dev-env setup to make it easier to install the same dependencies locally as get installed in the build (install locally now by running './dev-env/go-gets'). This isn't perfect, but it at least allows us to use atom with full linting and code completion.