add scaffolding for 'enable-scratch-volumes' global flag
This command will allow lc users to specify, on a per-repo basis, if they would like to utilize scratch-space optimizations. Note that currently all repose default to using scratch-space optimizations, however these optimizations introduce a layer of indirection that makes it hard to reason about what lc is doing. Once this config is implemented (in subsequent commits), this optimization will become an opt-in feature that defaults to off