break out scratch-volumes into an optional part of the template
This commit introduces a first class struct for lc templates. This struct allows the template provider to specify the exact scratch volumes that should be used if the 'enable-scratch-volume' flag is true. This isn't quite perfect yet as it requires the template provider know how to structure the composeYmlTmpl field with the correct variable for interpolating the scratch volumes at runtime.