perform smart cleanup in lc package
This ensures that containers created from a previous version of the docker image built during 'lc package' do not hang around when a new version of that image is built. Essentially this implements the narrative:

>I am running 'lc package' to produce a new version of my image, so I don't want any container created from the old image version to hang around after 'lc package' completes

This should fix issues engineers are hitting during 'blacbox-test' where they are testing with an old artifact.