add blackbox test for teardown of networks with -f
Relates to #5.

Compose does not offer a clean way to remove only networks; the only way to really remove them is to use `down`, but that would also remove our containers with the gc label. Compose also doesn't label networks (though hopefully it will soon: https://github.com/docker/compose/issues/3107) so we can't do the same type of thing we are doing for containers in `./command/teardown.go#removeContainersWithoutGcLabel`.

So until Compose offers a better way of finding/removing the networks it creates, just clarify the behavior in the blackbox-test (i.e., they are removed during `lc teardown -f`).