use public sbt image
Fixes #15

This switches the sbt image to use `paulcichonski/sbt`, which is still not official, but at least it is public and we can control it. Also added a strongly phrased recommendation to the docs saying that users should override this default image. This isn't ideal, but it is better than using an image that does not exist.