add support for dumping build logs
Fixes #43.

This commit adds support for the `build_logs_dir` config in `lc.yml`. If set, elsy will dump _all_ docker-compose service logs to the repo-relative folder set for the config, allocating a file per service. This does not filtering on the service logs (i.e., does not skip service if log data is empty), this lack of filtering is intentional as I find it to be simpler to filter un-needed data when debugging a build, than to be wondering why logs for certain services were not captured.