stop trying to pull repo's docker-image artifact during bootstrap
Fixes #78.

This refactors `CmdBootstrap` to filter down the services to pull by excluding any service using an `image:` declaration with the image from the repo's `docker-image-name`. This allows us to get rid of the brittle squelching of log messages by post-filtering the log stream. It also allows us to actualy fail `lc bootstrap` if images are not available, which is something that has caused confusion for a while.